requirements.txt
tensorflow==2.11.0
keras==2.11.0
opencv-python==4.5.5.64
imutils==0.5.4
streamlit==1.12.0
flask==2.1.2
pandas==1.4.2
numpy==1.22.4
Pillow==9.1.1